pyproject.toml
[project]
name = "gempilot"
version = "0.1.0"
description = "an LLM-powered command-line program capable of reading, updating, and running Python code using the Gemini API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
